// backend/src/enum/obsExportCsvHeader.enum.ts
export enum ObsExportCsvHeader {
  LocationId = "Location ID",
  LocationType = "Location Type",
  Latitude = "Latitude",
  Longitude = "Longitude",
  Elevation = "Elevation",
  ElevationUnit = "Elevation Unit",
  LocationGroup = "Location Group",
  FieldComment = "Field: Comment",
  FieldDeviceId = "Field: Device ID",
  FieldDeviceType = "Field: Device Type",
  CollectionMethod = "Collection Method",
  Medium = "Medium",
  Taxonomy = "Specimen: Taxonomy Element",
  DepthUpper = "Depth",
  DepthUnit = "Depth Unit",
  ObservedDateTime = "Observed DateTime",
  ObservedPropertyId = "Observed Property ID",
  CasNumber = "CAS Number",
  ResultValue = "Result Value",
  MethodDetectionLimit = "Lab: MDL",
  MethodReportingUnit = "Lab: MRL",
  ResultUnit = "Result Unit",
  DetectionCondition = "Lab: Detection Condition",
  LimitType = "Lab: Limit Type",
  CompositeStat = "EA_Observation Composite Stat",
  DataClassification = "Data Classification",
  AnalyzingAgency = "Lab: From Laboratory",
  AnalysisMethod = "Lab: Analysis Method",
  AnalyzedDateTime = "Analyzed DateTime",
  ResultStatus = "Result Status",
  ResultGrade = "Result Grade",
  ActivityName = "Activity Name",
  SpecimenName = "Lab: Specimen Name",
  LabQualityFlag = "Lab: Quality Flag",
  LabArrivalDateTime = "Lab: Received DateTime",
  LabPreparedDateTime = "Lab: Prepared DateTime",
  LabSampleId = "Lab: Sample ID",
  LabDilutionFactor = "Lab: Dilution Factor",
  LabComment = "Lab: Comment",
  LabBatchId = "EA_Lab Batch ID",
  QCType = "QC: Type",
  QCSourceActivityName = "QC: Source Sample ID",
  ValidationWarnings = "Validation Warnings",
  StandardsViolation = "Standards Violations",
  ObservationId = "Observation ID"
}